everything works

+++ src/controllers/songController.ts
@@ -6,8 +6,19 @@ import { User } from '../models/databaseModel';
 // Function to get all songs
 export async function getAllSongs(req: Request, res: Response): Promise<void> {
   try {
+    if (!req.user || !req.user.email) {
+      return handleResponse(res, 404, "error", "Not Found", "User account not found.");
+    }
+    const userEmail = req.user.email;
+     // check if the user exists
+     const existingUser = await User.findOne({ userEmail });
+    
+     if (!existingUser) {
+       // if user does not exist, return a not found error
+       return handleResponse(res, 404, "error", "Not Found", "User account not found.");
+     }
     const songs = await Song.find().exec();
-    res.json(songs);
+    res.status(200).json(songs);
   } catch (error) {
     res.status(500).json({ message: 'Error retrieving songs' });
   }
@@ -16,12 +27,23 @@ export async function getAllSongs(req: Request, res: Response): Promise<void> {
 // Function to get a song by id
 export async function getSongById(req: Request, res: Response): Promise<void> {
   try {
+    if (!req.user || !req.user.email) {
+      return handleResponse(res, 404, "error", "Not Found", "User account not found.");
+    }
+    const userEmail = req.user.email;
+     // check if the user exists
+     const existingUser = await User.findOne({ userEmail });
+    
+     if (!existingUser) {
+       // if user does not exist, return a not found error
+       return handleResponse(res, 404, "error", "Not Found", "User account not found.");
+     }
     const id = req.params.id;
     const song = await Song.findById(id).exec();
     if (!song) {
       res.status(404).json({ message: 'Song not found' });
     } else {
-      res.json(song);
+      res.status(200).json(song);
     }
   } catch (error) {
     res.status(500).json({ message: 'Error retrieving song' });
@@ -36,6 +58,7 @@ export async function addSong(req: Request, res: Response): Promise<void> {
       return handleResponse(res, 404, "error", "Not Found", "User account not found.");
     }
     const userEmail = req.user.email;
+    console.log(userEmail)
      // check if the user exists
      const existingUser = await User.findOne({ userEmail });
     
@@ -49,7 +72,7 @@ export async function addSong(req: Request, res: Response): Promise<void> {
      }
     const newSong = new Song(req.body);
     await newSong.save();
-    res.json(newSong);
+    res.status(200).json(newSong);
   } catch (error) {
     console.log(error)
     res.status(500).json({ message: 'Error adding song' });
@@ -80,7 +103,7 @@ export async function updateSong(request: Request, response: Response): Promise<
     if (!song) {
       response.status(404).json({ message: 'Song not found' });
     } else {
-      response.json(song);
+      response.status(200).json(song);
     }
   } catch (error) {
     response.status(500).json({ message: 'Error updating song' });
@@ -90,9 +113,20 @@ export async function updateSong(request: Request, response: Response): Promise<
 // Function to delete a song
 export async function deleteSong(req: Request, res: Response): Promise<void> {
   try {
+    if (!req.user || !req.user.email) {
+      return handleResponse(res, 404, "error", "Not Found", "User account not found.");
+    }
+    const userEmail = req.user.email;
+     // check if the user exists
+     const existingUser = await User.findOne({ userEmail });
+    
+     if (!existingUser) {
+       // if user does not exist, return a not found error
+       return handleResponse(res, 404, "error", "Not Found", "User account not found.");
+     }
     const id = req.params.id;
     await Song.findByIdAndDelete(id).exec();
-    res.json({ message: 'Song deleted successfully' });
+    res.status(200).json({ message: 'Song deleted successfully' });
   } catch (error) {
     res.status(500).json({ message: 'Error deleting song' });
   }
@@ -102,9 +136,20 @@ export async function deleteSong(req: Request, res: Response): Promise<void> {
 // Function to get songs by artist
 export async function getSongsByArtist(req: Request, res: Response): Promise<void> {
   try {
-    const artist = req.query.artist;
-    const songs = await Song.find({ artist: { $regex: artist, $options: 'i' } }).exec();
-    res.json(songs);
+    if (!req.user || !req.user.email) {
+      return handleResponse(res, 404, "error", "Not Found", "User account not found.");
+    }
+    const userEmail = req.user.email;
+     // check if the user exists
+     const existingUser = await User.findOne({ userEmail });
+    
+     if (!existingUser) {
+       // if user does not exist, return a not found error
+       return handleResponse(res, 404, "error", "Not Found", "User account not found.");
+     }
+    const artist = req.query.q;
+    const songs = await Song.find({songArtistId: artist }).exec();
+    res.status(200).json(songs);
   } catch (error) {
     res.status(500).json({ message: 'Error searching songs by artist' });
   }
@@ -113,9 +158,20 @@ export async function getSongsByArtist(req: Request, res: Response): Promise<voi
 // Function to get songs by genre
 export async function getSongsByGenre(req: Request, res: Response): Promise<void> {
   try {
-    const genre = req.query.genre;
+    if (!req.user || !req.user.email) {
+      return handleResponse(res, 404, "error", "Not Found", "User account not found.");
+    }
+    const userEmail = req.user.email;
+     // check if the user exists
+     const existingUser = await User.findOne({ userEmail });
+    
+     if (!existingUser) {
+       // if user does not exist, return a not found error
+       return handleResponse(res, 404, "error", "Not Found", "User account not found.");
+     }
+    const genre = req.query.q;
     const songs = await Song.find({ genre: { $regex: genre, $options: 'i' } }).exec();
-    res.json(songs);
+    res.status(200).json(songs);
   } catch (error) {
     res.status(500).json({ message: 'Error searching songs by genre' });
   }
@@ -124,9 +180,20 @@ export async function getSongsByGenre(req: Request, res: Response): Promise<void
 // Function to get songs by date
 export async function getSongsByDate(req: Request, res: Response): Promise<void> {
   try {
-    const recordedDate = req.query.recordedDate;
-    const songs = await Song.find({ recordedDate: recordedDate }).exec();
-    res.json(songs);
+    if (!req.user || !req.user.email) {
+      return handleResponse(res, 404, "error", "Not Found", "User account not found.");
+    }
+    const userEmail = req.user.email;
+     // check if the user exists
+     const existingUser = await User.findOne({ userEmail });
+    
+     if (!existingUser) {
+       // if user does not exist, return a not found error
+       return handleResponse(res, 404, "error", "Not Found", "User account not found.");
+     }
+    const recordedDate = req.query.q;
+    const songs = await Song.find({ recordedDate }).exec();
+    res.status(200).json(songs);
   } catch (error) {
     res.status(500).json({ message: 'Error searching songs by date' });
   }
@@ -134,7 +201,18 @@ export async function getSongsByDate(req: Request, res: Response): Promise<void>
 //Function to get songs by partial search
 export async function getSongsByLyrics(req: Request, res: Response): Promise<void> {
   try {
-    const lyrics = req.query.lyrics; // Assuming the lyrics are passed as a query parameter
+    if (!req.user || !req.user.email) {
+      return handleResponse(res, 404, "error", "Not Found", "User account not found.");
+    }
+    const userEmail = req.user.email;
+     // check if the user exists
+     const existingUser = await User.findOne({ userEmail });
+    
+     if (!existingUser) {
+       // if user does not exist, return a not found error
+       return handleResponse(res, 404, "error", "Not Found", "User account not found.");
+     }
+    const lyrics = req.query.q; // Assuming the lyrics are passed as a query parameter
     const songs = await Song.find({
       lyrics: { $regex: lyrics, $options: 'i' } // Use regex for partial match and ignore case
     }).exec();
@@ -142,7 +220,7 @@ export async function getSongsByLyrics(req: Request, res: Response): Promise<voi
     if (songs.length === 0) {
       res.status(404).json({ message: 'No songs found with those lyrics' });
     } else {
-      res.json(songs);
+      res.status(200).json(songs);
     }
   } catch (error) {
     res.status(500).json({ message: 'Error searching songs by partial lyrics' });
